`@remotion/install-whisper-cpp`: Make it install with Whisper.cpp on Windows
