+++ packages/install-whisper-cpp/src/index.ts
@@ -1,9 +1,6 @@
 export {Caption, convertToCaptions} from './convert-to-captions';
-export {
-	OnProgress,
-	WhisperModel,
-	downloadWhisperModel,
-} from './download-whisper-model';
+export type {OnProgress} from './download';
+export {WhisperModel, downloadWhisperModel} from './download-whisper-model';
 export {installWhisperCpp} from './install-whisper-cpp';
 export type {Language} from './languages';
 export {TranscriptionJson, transcribe} from './transcribe';
