+++ packages/install-whisper-cpp/src/install-whisper-cpp.ts
@@ -1,11 +1,9 @@
 import type {StdioOptions} from 'child_process';
 import fs, {existsSync, rmSync} from 'fs';
 import {execSync} from 'node:child_process';
-import {Readable} from 'node:stream';
-import {finished} from 'node:stream/promises';
-import type {ReadableStream} from 'node:stream/web';
 import os from 'os';
 import path from 'path';
+import {downloadFile} from './download';
 
 const getIsSemVer = (str: string) => {
 	return /^[\d]{1}\.[\d]{1,2}\.+/.test(str);
@@ -33,14 +31,12 @@ const installForWindows = async ({
 	const filePath = path.join(process.cwd(), 'whisper-bin-x64.zip');
 	const fileStream = fs.createWriteStream(filePath);
 
-	const {body} = await fetch(url);
-	if (body === null) {
-		throw new Error('Failed to download whisper binary');
-	}
-
-	await finished(
-		Readable.fromWeb(body as unknown as ReadableStream).pipe(fileStream),
-	);
+	await downloadFile({
+		fileStream,
+		printOutput,
+		url,
+		onProgress: undefined,
+	});
 
 	execSync(`Expand-Archive -Force ${filePath} ${to}`, {
 		shell: 'powershell',
